[1208]첫번째 커밋
