+++ workspace/rushwash/src/main/java/com/rushwash/app/payment/controller/CardRegistController.java
@@ -1,4 +1,4 @@
-package com.rushwash.app.payment;
+package com.rushwash.app.payment.controller;
 
 import java.io.IOException;
 
@@ -8,10 +8,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/payment/laundry-form")
-public class FormWriteController extends HttpServlet{
+@WebServlet("/payment/card-regist")
+public class CardRegistController extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		req.getRequestDispatcher("/WEB-INF/views/user/payment/laundry_form.jsp").forward(req, resp);
+		req.getRequestDispatcher("/WEB-INF/views/user/payment/card_regist.jsp").forward(req, resp);
 	}
 }
