+++ workspace/rushwash/src/main/java/com/rushwash/app/payment/service/PaymentService.java
@@ -1,0 +1,5 @@
+package com.rushwash.app.payment.service;
+
+public class PaymentService {
+
+}
